pyproject.toml
[tool.poetry]
name = "hostaway-interview-data-engineer"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python =  ">=3.11,<3.12"
pandas = "^2.2.3"
apache-airflow = "2.7.3"
apache-airflow-providers-postgres = "^5.7.1"
dbt-core = "==1.9.3"
dbt-postgres = "==1.9.0"
